Color Theme for Graphics

color.theme() returns an object of class "color.theme" that provides two types of color functions.

Usage

color.theme(colors, type = c("sequential", "qualitative", "diverging"), ...)

Arguments

colors

one of the following: a color theme name such as "Viridis" with the optional suffix "_r" for color themes in reverse order ("Viridis_r"), a character vector of color names, a palette function, or a ramp function to be used to create a color theme.

type

a character string specifying the type of the color theme: One of "sequential", "qualitative" or "diverging".

...

optional arguments to be passed to palette or ramp functions.

Value

color.theme() returns a "color.theme" object containing following components:

ramp

the function that takes a numeric vector x of the values within [0, 1] and returns a color name vector.

palette

the function that takes an integer n and returns a color name vector of length n.

type

the type of the color theme; "sequential", "diverging" or "qualitative".

Details

"color.theme" objects is a container of the two types of color functions: palette(n) returns a color name vector of length n, and ramp(x) returns color names for each values of x within [0, 1]. Some color themes are "qualitative" and do not contain ramp() function.

Examples

color.theme("Zissou 1")$palette(5L)
#> [1] "#3B99B1" "#7CBA96" "#EACB2B" "#E78F0A" "#F5191C"
color.theme("Zissou 1")$ramp(seq.int(0, 1, 1/4))
#> [1] "#3A99B1" "#7CB997" "#D3BC5E" "#E78D08" "#F4191B"
color.theme("Zissou 1_r")$ramp(seq.int(1, 0, -1/4))
#> [1] "#3A99B1" "#7CB997" "#D3BC5E" "#E78D08" "#F4191B"
color.theme("Viridis", alpha = .5)$palette(5L)
#> [1] "#4B005580" "#00588B80" "#009B9580" "#53CC6780" "#FDE33380"
color.theme("Viridis_r", alpha = .2)$palette(5L)
#> [1] "#FDE33333" "#53CC6733" "#009B9533" "#00588B33" "#4B005533"
color.theme("midr")$palette(5L)
#> [1] "#2E7999" "#9CBACB" "#FFFFFF" "#C18DA4" "#7E1952"