Calculate ICE of MID Models

mid.conditional() creates an object to draw ICE curves of a MID model.

Usage

mid.conditional(
  object,
  variable,
  data,
  keep.effects = TRUE,
  n.samples = 100L,
  max.nrow = 100000L,
  type = c("response", "link")
)

# S3 method for class 'mid.conditional'
print(x, ...)

Arguments

object

a "mid" object.

variable

a character string or expression specifying the variable for the ICE calculation.

data

a data frame containing observations for which ICE values are calculated.

keep.effects

logical. If TRUE, the effects of component functions are stored in the output object.

n.samples

integer. The number of sample points for the calculation.

max.nrow

an integer specifying the maximum number of rows of the output data frames.

type

the type of prediction required. The default is "response". "link" is possible if the MID model uses a link function.

x

a "mid.conditional" object to be printed.

...

additional parameters to be passed to print.default() to print the sample point vector.

Value

mid.conditional() returns an object of class "mid.conditional" with the following components:

terms

the character vector of relevant terms.

observed

the data frame of the actual observations and the corresponding predictions.

conditional

the data frame of the hypothetical observations and the corresponding predictions.

values

the sample points of the variable.

Details

mid.conditional() obtains predictions for hypothetical observations from a MID model and returns a "mid.conditional" object. The graphing functions ggmid() and plot() can be used to generate the ICE curve plots.

Examples

data(airquality, package = "datasets")
mid <- interpret(Ozone ~ .^2, airquality, lambda = 1)
#> 'model' is not passed: the response variable in the data is used
mc <- mid.conditional(mid, "Wind", airquality)
mc
#> 
#> Individual Conditional Expectation for 153 Observations
#> 
#> Variable: Wind
#> 
#> Sample Points:
#>   [1]  2.300000  2.485859  2.671717  2.857576  3.043434  3.229293  3.415152
#>   [8]  3.601010  3.786869  3.972727  4.158586  4.344444  4.530303  4.716162
#>  [15]  4.902020  5.087879  5.273737  5.459596  5.645455  5.831313  6.017172
#>  [22]  6.203030  6.388889  6.574747  6.760606  6.946465  7.132323  7.318182
#>  [29]  7.504040  7.689899  7.875758  8.061616  8.247475  8.433333  8.619192
#>  [36]  8.805051  8.990909  9.176768  9.362626  9.548485  9.734343  9.920202
#>  [43] 10.106061 10.291919 10.477778 10.663636 10.849495 11.035354 11.221212
#>  [50] 11.407071 11.592929 11.778788 11.964646 12.150505 12.336364 12.522222
#>  [57] 12.708081 12.893939 13.079798 13.265657 13.451515 13.637374 13.823232
#>  [64] 14.009091 14.194949 14.380808 14.566667 14.752525 14.938384 15.124242
#>  [71] 15.310101 15.495960 15.681818 15.867677 16.053535 16.239394 16.425253
#>  [78] 16.611111 16.796970 16.982828 17.168687 17.354545 17.540404 17.726263
#>  [85] 17.912121 18.097980 18.283838 18.469697 18.655556 18.841414 19.027273
#>  [92] 19.213131 19.398990 19.584848 19.770707 19.956566 20.142424 20.328283
#>  [99] 20.514141 20.700000