midr: Create an Interpretable Surrogate of Black-Box ML Models

Construct a globally interpretable surrogate of a black-box ML model by decomposing the prediction function of the target model into a set of component functions including the intercept, the main effects and the second-order interactions.

See also

Author

Maintainer: Ryoichi Asasihba ryoichi.asashiba@gmail.com

Authors:

Hirokazu Iwasawa

Other contributors:

Reiji Kozuma [contributor]