Weighted Loss Functions

weighted.rmse(), weighted.mae() and weighted.medae() compute the loss from a weighted vector of prediction errors.

Usage

weighted.rmse(x, y = NULL, w = NULL, ..., na.rm = FALSE)

weighted.mae(x, y = NULL, w = NULL, ..., na.rm = FALSE)

weighted.medae(x, y = NULL, w = NULL, ..., na.rm = FALSE)

Arguments

x

a numeric vector of errors.

y

an optional numeric vector. If passed, the loss is calculated for the differences between x and y.

w

a numeric vector of sample weights for each value in x.

...

optional parameters.

na.rm

logical. If TRUE, any NA and NaNs are removed from x before the calculation.

Value

weighted.rmse() (root mean squared error), weighted.mae() (mean absolute error) and weighted.medae (median absolute error) returns a single numeric value.

Details

weighted.rmse() returns the root mean square error, weighted.mae() returns the mean absolute error, and weighted.medae() returns the median absolute error for a weighted vector.

Examples

weighted.rmse(x = c(0, 10), y = c(0, 0), w = c(99, 1))
#> [1] 1
weighted.mae(x = c(0, 10), y = c(0, 0), w = c(99, 1))
#> [1] 0.1
weighted.medae(x = c(0, 10), y = c(0, 0), w = c(99, 1))
#> [1] 0