Calculate MID Importance

mid.importance() calculates the MID importance of a fitted MID model.

Usage

mid.importance(object, data = NULL, weights = NULL, sort = TRUE, measure = 1L)

Arguments

object

a "mid" object.

data

a data frame containing the observations to be used to calculate the MID importance. If NULL, the fitted.matrix of the MID model is used. If the data has only one observation, the output has the special class "mid.breakdown".

weights

an optional numeric vector of sample weights.

sort

logical. If TRUE, the output data frame is sorted by MID importance.

measure

an integer specifying the measure of the MID importance. Possible alternatives are 1 for the mean absolute effect, 2 for the root mean square effect, and 3 for the median absolute effect.

Value

mid.importance returns a data frame of the class "mid.importance".

Details

mid.importance() returns an object of class "mid.importance". The MID importance is defined for each component function of a MID model as the mean absolute effect in the given data.

Examples

data(airquality, package = "datasets")
mid <- interpret(Ozone ~ .^2, airquality, lambda = 1)
#> 'model' is not passed: the response variable in the data is used
imp <- mid.importance(mid)
imp
#>             term importance degree
#> 1           Temp 13.9409625      1
#> 2           Wind 10.5054026      1
#> 3        Solar.R  5.5745798      1
#> 4            Day  4.3927693      1
#> 5          Month  2.2799537      1
#> 6   Solar.R:Wind  0.4494962      2
#> 7  Solar.R:Month  0.3762348      2
#> 8       Temp:Day  0.3714793      2
#> 9     Wind:Month  0.3690533      2
#> 10      Wind:Day  0.3654715      2
#> 11     Month:Day  0.3226006      2
#> 12     Wind:Temp  0.2805991      2
#> 13   Solar.R:Day  0.2664565      2
#> 14  Solar.R:Temp  0.2346434      2
#> 15    Temp:Month  0.1207601      2